+++ copasi/compareExpressions/compare_utilities.cpp
@@ -1,9 +1,9 @@
 // Begin CVS Header
 //   $Source: /Volumes/Home/Users/shoops/cvs/copasi_dev/copasi/compareExpressions/compare_utilities.cpp,v $
-//   $Revision: 1.12 $
+//   $Revision: 1.12.2.1 $
 //   $Name:  $
-//   $Author: gauges $
-//   $Date: 2010/04/19 12:53:49 $
+//   $Author: shoops $
+//   $Date: 2011/01/10 17:00:27 $
 // End CVS Header
 
 // Copyright (C) 2010 by Pedro Mendes, Virginia Tech Intellectual
@@ -1066,7 +1066,7 @@ CEvaluationNode* expand_function_call(const CEvaluationNodeCall* pCall, CFunctio
       // create the mapping
       assert(pFunctionDefinition != NULL);
       const CFunctionParameters& functionParameters = pFunctionDefinition->getVariables();
-      unsigned int i, iMax = functionParameters.size();
+      size_t i, iMax = functionParameters.size();
       // map the first function argument to the first child in the call etc.
       std::map<std::string, const CEvaluationNode*> argumentMap;
       i = 0;
